fix: 튜토리얼 및 UI 관련 버그 수정
- TutorialManager: IsTutorial 프로퍼티를 캔버스 존재 여부로 판단하도록 변경
- InGameSynergyUI: 시너지 게이지 fillAmount 계산 오류 수정 (정수 나눗셈 → 실수 나눗셈)
- InGameBottomUI: 튜토리얼 진행 중 새 캐릭터 아이템에 TutorialTarget 등록 추가
- SceneUILayerManager: 씬 전환 시 이미 파괴된 UI 스킵 처리 추가
- .gitignore: Claude 관련 파일 제외 규칙 추가

+++ Assets/_Project/Scripts/UI/Common/InGameSynergyUI.cs
@@ -65,7 +65,7 @@ namespace CookApps.AutoBattler
 
             var lastSynergyData = SpecDataManager.Instance.GetSpecSynergyList(synergyType).Last();
 
-            _gradeGuageImage.fillAmount = (float)(data.grade / lastSynergyData.grade);
+            _gradeGuageImage.fillAmount = (float)data.grade / (float)lastSynergyData.grade;
             _gradeGuageImage.color = color;
 
             //  .color = (isColorWhite) ? color : Color.white;
